operators/ebpf: Fix crash when spec doesn't contain types
Compiling an empty progra.bpf.c file will produce an eBPF layer with a
collection spec without types, this causes a crash when trying to run
such a gadget.

This commit add an additional check to ensure the collection spec
contains types, and also adds a test case to cover this scenario.

Fixes: d28c76d0ef85 ("pkg/operators/ebpf: ebpf layer operator")

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>