pkg: Add safeset package.
Before this commit, we used the private variable safeSet available in
encoding/json through linkname.
But, since golang 1.23, linkname is not authorized by default [1].

So, this commit introduces the safeset package which defines the SafeSet map.
This map has the exact same value as encoding/json.safeSet [2].
It is also safe to add it in our sources, as this variable was not modified
upstream since its introduction in 2016 [3].

This permits the whole project to be compiled using golang 1.23.

Signed-off-by: Francis Laniel <flaniel@linux.microsoft.com>
[1]: https://github.com/golang/go/issues/67401
[2]: https://cs.opensource.google/go/go/+/release-branch.go1.23:src/encoding/json/tables.go;l=15-112
[3]: https://cs.opensource.google/go/go/+/ed8f207940c8787d344664a43071b235e2ce5c68