operators/cli: Improve logic to print header
Printing the header when using the columns mode is a simple but complicated
task as there are multiple things to take into account.

1. It should only be printed once for gadgets that dump all the
information at once. Like snapshotters.
2. It should be printed right away for gadgets that collect and print the
information later on (like toppers).

This commit is a "best effort" to fix that, probably there are corner
cases not covered yet.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>