Makefile: Introduce target to run go mod tidy
We have different golang modules in the repository, un fortunately there
is a not an easy way to update all of them. This commit introduces a new
target to the Makefile that runs `go mod tidy` in all the modules.
This is useful to update all modules when the main one is updated
by dependabot.

Signed-off-by: Mauricio Vásquez <mauriciov@microsoft.com>